requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
Pillow>=8.2.0
pytesseract>=0.3.8
lxml>=4.6.3
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5